% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestorm-package.R
\docType{package}
\name{firestorm-package}
\alias{firestorm}
\alias{firestorm-package}
\title{firestorm: Reverse Proxy and Load Balancing for Fiery}
\description{
Provides facilities for setting up fiery apps as a reverse proxy
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
