% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestorm-package.R
\docType{package}
\name{firestorm-package}
\alias{firestorm}
\alias{firestorm-package}
\title{firestorm: Reverse Proxy and Load Balancing for Fiery}
\description{
Provides plugins for setting up fiery apps as a reverse proxy. This allows you to use a fiery server as a front for multiple services or even work as a load-balancer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/firestorm}
  \item Report bugs at \url{https://github.com/thomasp85/firestorm/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
